final lesson 5

+++ MetricsAgent/Models/CpuMetric.cs
@@ -1,10 +1,10 @@
-﻿namespace MetricsAgent.Models
+﻿ namespace MetricsAgent.Models
 {
     public class CpuMetric
     {
         public int Id { get; set; }
         public int Value { get; set; }
-        public int Time { get; set; }
+        public long Time { get; set; }
 
     }
 }
